destroy function on tag&category

+++ app/Http/Controllers/TagsController.php
@@ -68,6 +68,13 @@ class TagsController extends Controller {
    * Remove the specified resource from storage.
    */
   public function destroy(Tag $tag) {
+
+    if ($tag->posts->count() > 0) {
+      session()->flash('error', 'Tag cannot be deleted because it is associated to some posts.');
+
+      return redirect()->back();
+    }
+
     $tag->delete();
     session()->flash('success', 'Tag deleted successfully.');
     return redirect(route('tags.index'));
